move coin specific operation related logic in coin module
