+++ libs/coin-evm/package.json
@@ -66,6 +66,7 @@
     "imurmurhash": "^0.1.4",
     "invariant": "^2.2.2",
     "lodash": "^4.17.21",
+    "prando": "^6.0.1",
     "react-native-get-random-values": "^1.8.0",
     "rxjs": "^7.8.1"
   },
