+++ libs/ledger-live-common/scripts/sync-families-dispatch.mjs
@@ -16,6 +16,7 @@ const targets = [
   "exchange.ts",
   "platformAdapter.ts",
   "walletApiAdapter.ts",
+  "operation.ts",
 ];
 
 // Coins using coin-framework
